requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.41.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
PyAudio==0.2.13
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soundfile==0.12.1
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4